refactoring IdentitiesSection
