+++ src/@types/identitiesSectionContent.ts
@@ -1,0 +1,8 @@
+import { Identity } from "@lib/@types";
+
+export type IdentitiesSectionContent = {
+  title: string;
+  noIdentityMessage: string;
+  addNewIdentityMessage?: string;
+  identitiesList: Identity[];
+};
